requirements.txt
torch>=2.0.1
accelerate
appdirs
loralib
bitsandbytes
black
black[jupyter]
datasets
fire
peft
transformers>=4.31.0
sentencepiece
py7zr
scipy
optimum
wandb
hydra-core>=1.3.2
openai-whisper
wandb
soundfile
evaluate
transformers
datasets
sacrebleu
jiwer
librosa
unbabel-comet
nltk
openpyxl
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
debugpy